+++ android/services/mongodb-remote-perftests/src/androidTest/java/com/mongodb/stitch/android/services/mongodb/remote/SyncPerformanceIntTestsHarness.kt
@@ -189,6 +189,7 @@ class SyncPerformanceIntTestsHarness : BaseStitchAndroidIntTest() {
                             .append("exception", e.localizedMessage)
                             .append("stacktrace", e.stackTrace.map { it.toString() }))
             ))
+            throw e
         }
     }
 }
