+++ android/services/mongodb-remote-perftests/src/androidTest/java/com/mongodb/stitch/android/services/mongodb/remote/SyncPerformanceTestContext.kt
@@ -116,7 +116,7 @@ class SyncPerformanceTestContext(
         // TODO: The test should be set up in such a way that each setup sets up an embedded mongo
         // instance. Until we do that, we can't actually properly tear down the embedded instance
         // or close the data synchronizer on each test iteration.
-        
+
         // if (::testMongoClient.isInitialized) {
         //    (testMongoClient as RemoteMongoClientImpl).dataSynchronizer.close()
         //    AndroidEmbeddedMongoClientFactory.getInstance().close()
