+++ android/services/mongodb-remote-perftests/src/androidTest/java/com/mongodb/stitch/android/services/mongodb/remote/SyncPerformanceTests.kt
@@ -26,19 +26,6 @@ class SyncPerformanceTests {
 
     private var testHarness = SyncPerformanceIntTestsHarness()
 
-    private fun getDocuments(
-        numberOfDocs: Int,
-        sizeOfDocsInBytes: Int
-    ): List<Document>? {
-        val array: List<Byte> = (0 until sizeOfDocsInBytes).map { 0.toByte() }
-        return (0 until numberOfDocs).map {
-            Document(mapOf(
-                "_id" to ObjectId(),
-                "bin" to Binary(array.toByteArray())
-            ))
-        }
-    }
-
     @Before
     fun setup() {
         this.testHarness.setup()
@@ -49,6 +36,24 @@ class SyncPerformanceTests {
         this.testHarness.teardown()
     }
 
+    // Tests for L2R-only scenarios
+
+    @Test
+    fun testL2ROnlyInitialSync() {
+        SyncL2ROnlyPerformanceTestDefinitions.testInitialSync(this.testHarness, runId)
+    }
+
+    @Test
+    fun testL2ROnlyDisconnectReconnect() {
+        SyncL2ROnlyPerformanceTestDefinitions.testDisconnectReconnect(this.testHarness, runId)
+    }
+
+    @Test
+    fun testL2ROnlySyncPass() {
+        SyncL2ROnlyPerformanceTestDefinitions.testSyncPass(this.testHarness, runId)
+    }
+
+    // Placeholder tests until the performance testing suite is in place
     @Test
     fun testInitialSync() {
         val testParams = TestParams(
@@ -81,7 +86,8 @@ class SyncPerformanceTests {
             numDocs = intArrayOf(10),
             numIters = 3,
             numOutliersEachSide = 0,
-            outputToStitch = outputToStitch
+            outputToStitch = outputToStitch,
+            stitchHostName = "https://stitch.mongodb.com"
         )
 
         testHarness.runPerformanceTestWithParams(testParams, { ctx, docSize, numDocs ->
@@ -90,5 +96,18 @@ class SyncPerformanceTests {
             Tasks.await(ctx.testColl.insertMany(documents))
             assertEquals(Tasks.await(ctx.testColl.count()), numDocs.toLong())
         })
+    }
+
+    private fun getDocuments(
+        numberOfDocs: Int,
+        sizeOfDocsInBytes: Int
+    ): List<Document>? {
+        val array: List<Byte> = (0 until sizeOfDocsInBytes).map { 0.toByte() }
+        return (0 until numberOfDocs).map {
+            Document(mapOf(
+                    "_id" to ObjectId(),
+                    "bin" to Binary(array.toByteArray())
+            ))
+        }
     }
 }
